all discussion titles show up
