+++ src/js/discussions/getDiscussionsReducer.js
@@ -1,3 +1,13 @@
+const discussion = (state = {}, action) => {
+  if(action.type === "CREATE_DISCUSSION_SUCCESS") {
+    return {
+      id: action.id,
+      inputStr: action.inputStr
+    }
+  }
+  return state;
+}
+
 const discussionsGet = (state = {isFetching: false, discussions: [], error: null}, action) => {
   if(action.type === 'GET_DISCUSSIONS_SUCCESS') {
     console.log('action discussions', action.discussions)
@@ -10,6 +20,13 @@ const discussionsGet = (state = {isFetching: false, discussions: [], error: null
         discussions: action.discussions
       }
     )
+  }
+  if(action.type === "CREATE_DISCUSSION_SUCCESS") {
+    console.log('discussion made!!!', action)
+    return [
+      discussion(undefined, action),
+      ...state
+    ]
   }
   return state
 }
